refactor: ProblemSolveContainer 분리
