+++ frontend/src/components/Common/HStack/index.ts
@@ -1,1 +1,2 @@
 export { HStack } from './HStack';
+export type { Props as HStackProps } from './HStack';
