+++ frontend/src/pages/CompetitionPage.tsx
@@ -1,33 +1,23 @@
 import { css } from '@style/css';
 
-import { useContext, useMemo, useState } from 'react';
+import { useMemo, useState } from 'react';
 import { useNavigate, useParams } from 'react-router-dom';
 
-import { Button, HStack, Modal, Space, VStack } from '@/components/Common';
+import { Button, Space } from '@/components/Common';
 import BreadCrumb from '@/components/Common/BreadCrumb';
 import Logo from '@/components/Common/Logo';
 import { SocketProvider } from '@/components/Common/Socket/SocketProvider';
 import CompetitionHeader from '@/components/Competition/CompetitionHeader';
 import CompetitionProblemSelector from '@/components/Competition/CompetitionProblemSelector';
 import DashboardModal from '@/components/Dashboard/DashboardModal';
-import Editor from '@/components/Editor/Editor';
 import { PageLayout } from '@/components/Layout/PageLayout';
 import { ProblemHeader } from '@/components/Problem/ProblemHeader';
-import ProblemViewer from '@/components/Problem/ProblemViewer';
-import { SimulationExecButton } from '@/components/Simulation/SimulationExecButton';
-import { SimulationInputModal } from '@/components/Simulation/SimulationInputModal';
-import { SimulationResultList } from '@/components/Simulation/SimulationResultList';
+import { ProblemSolveContainer } from '@/components/Problem/ProblemSolveContainer';
 import SocketTimer from '@/components/SocketTimer';
-import { SubmissionButton } from '@/components/Submission/SubmissionButton';
-import { SubmissionResult } from '@/components/Submission/SubmissionResult';
 import { ROUTE, SITE } from '@/constants';
 import { useCompetition } from '@/hooks/competition';
-import { useUserCode } from '@/hooks/editor/useUserCode';
-import useAuth from '@/hooks/login/useAuth';
 import { useCompetitionProblem } from '@/hooks/problem';
 import { useCompetitionProblemList } from '@/hooks/problem/useCompetitionProblemList';
-import { SimulationInput, useSimulation } from '@/hooks/simulation';
-import * as customLocalStorage from '@/utils/localStorage';
 
 export default function CompetitionPage() {
   const { id } = useParams<{ id: string }>();
@@ -53,40 +43,8 @@ export default function CompetitionPage() {
   }, [problemList, currentProblemIndex]);
 
   const { problem } = useCompetitionProblem(currentProblem?.id ?? -1);
-  const simulation = useSimulation(problem.testcases);
-
-  const { email } = useAuth();
-
-  const { code, setCode } = useUserCode({
-    userId: email,
-    problem,
-    competitionId,
-    currentProblemIndex,
-    save: customLocalStorage.save,
-  });
-
-  const handleChangeCode = (newCode: string) => {
-    setCode(newCode);
-  };
 
   const problemIds = problemList.map((problem) => problem.id);
-
-  const handleSimulate = () => {
-    simulation.run(code);
-  };
-
-  const handleSimulationCancel = () => {
-    simulation.cancel();
-  };
-  const modal = useContext(Modal.Context);
-
-  const handleSaveSimulationInputs = (simulationInputs: SimulationInput[]) => {
-    simulation.changeInputs(simulationInputs);
-  };
-
-  function handleOpenModal() {
-    modal.open();
-  }
 
   function handleTimeout() {
     navigate(`${ROUTE.DASHBOARD}/${competitionId}`);
@@ -120,39 +78,23 @@ export default function CompetitionPage() {
               onChangeProblemIndex={setCurrentProblemIndex}
             />
           </aside>
-          <HStack className={css({ height: '100%' })}>
-            <VStack className={problemSolveContainerStyle}>
-              <ProblemViewer className={problemStyle} content={problem.content}></ProblemViewer>
-              <HStack className={solutionStyle}>
-                <Editor height="500px" code={code} onChangeCode={handleChangeCode}></Editor>
-                <section>
-                  <SimulationResultList resultList={simulation.results}></SimulationResultList>
-                  <SubmissionResult></SubmissionResult>
-                </section>
-              </HStack>
-            </VStack>
-            <VStack as="footer" className={footerStyle}>
-              <Button onClick={handleOpenModal}>테스트 케이스 추가하기</Button>
-              <Space></Space>
-              <SimulationExecButton
-                isRunning={simulation.isRunning}
-                onExec={handleSimulate}
-                onCancel={handleSimulationCancel}
-              />
-              <SubmissionButton
-                code={code}
-                problemId={currentProblem?.id}
-                competitionId={competitionId}
-              ></SubmissionButton>
-            </VStack>
-          </HStack>
+          <ProblemSolveContainer
+            competitionId={0}
+            problem={{
+              id: 0,
+              title: '',
+              timeLimit: 0,
+              memoryLimit: 0,
+              content: '',
+              solutionCode: '',
+              testcases: [],
+              createdAt: '',
+            }}
+          ></ProblemSolveContainer>
         </div>
         <DashboardModal isOpen={isDashboardModalOpen} onClose={closeDashboardModal} />
       </SocketProvider>
-      <SimulationInputModal
-        simulationInputs={simulation.inputs}
-        onSave={handleSaveSimulationInputs}
-      ></SimulationInputModal>
+      <DashboardModal isOpen={isDashboardModalOpen} onClose={closeDashboardModal} />
     </PageLayout>
   );
 }
@@ -168,20 +110,6 @@ const competitionStyle = css({
   flexGrow: '1',
   overflow: 'hidden',
 });
-const problemSolveContainerStyle = css({
-  height: 'calc(100% - 4rem)',
-  width: 'full',
-});
-const problemStyle = css({
-  width: '1/2',
-  height: '100%',
-});
-const solutionStyle = css({
-  width: '1/2',
-  height: '100%',
-  alignItems: 'stretch',
-  overflow: 'auto',
-});
 
 const padVerticalStyle = css({
   paddingX: '1rem',
@@ -195,13 +123,4 @@ const asideStyle = css({
   padding: '0.5rem',
   width: '5rem',
   height: '100%',
-});
-
-const footerStyle = css({
-  height: '4rem',
-  width: '100%',
-  paddingX: '1rem',
-  gap: '0.5rem',
-  borderTop: '1px solid',
-  border: 'border',
 });
